cipher/rekey: fix Cipher REKEY operation
As per the Noise Protocol specification, the default REKEY should use the
first 32 bytes returned by ENCRYPT acting on 32 bytes of zeros with the
current key, a reserved nonce, and no associated data.

There were two problems with the existing implementation:

	1.  It was rekeying with the encrypted result plus the
	    authentication data, causing Cipher to fail on the next
	    encrypt/decrypt call due to an invalid key length.

	2.  It was resetting the nonce, which is explicitly disallowed by
	    specification.

This change clamps the rekey to the ciphertext sans the authentication data
and directly writes the key rather than calling initialize_key, which would
reset the nonce.

Signed-off-by: Derrick Pallas <derrick@pallas.us>